--- app/javascript/angular/code/aircasting.js
@@ -1,2 +1,0 @@
-angular.module("google", []);
-angular.module("aircasting", ["google"]);
