--- app/javascript/angular/code/controllers/sessions_map_ctrl.js
@@ -1,5 +1,0 @@
-import SessionsMapCtrl from "./_sessions_map_ctrl";
-
-angular
-  .module("aircasting")
-  .controller("SessionsMapCtrl", ["$scope", "$window", SessionsMapCtrl]);
