+++ app/javascript/elm/src/Main.elm
@@ -1203,8 +1203,7 @@ viewHeatMapInput text_ value_ sensorUnit toMsg =
 viewSessionsOrSelectedSession : Page -> WebData SelectedSession -> Int -> List Session -> WebData HeatMapThresholds -> Path -> Popup -> EmailForm -> Html Msg
 viewSessionsOrSelectedSession page selectedSession fetchableSessionsCount sessions heatMapThresholds linkIcon popup emailForm =
     div
-        [ attribute "ng-controller" "SessionsMapCtrl"
-        ]
+        []
         [ div [ class "sessions" ]
             [ case selectedSession of
                 NotAsked ->
