+++ package.json
@@ -10,7 +10,6 @@
   "dependencies": {
     "@google/markerclustererplus": "^2.1.11",
     "@rails/webpacker": "^4.0.7",
-    "angular": "1.2.29",
     "clipboard": "^2.0.4",
     "core-js": "3",
     "elm": "^0.19.0-no-deps",
@@ -49,7 +48,7 @@
     ],
     "all": true,
     "include": [
-      "app/javascript/angular/code/**/*.js"
+      "app/javascript/javascript/**/*.js"
     ],
     "sourceMap": false,
     "instrument": false
